requirements.txt
asgiref==3.5.2
Django==4.1.4
djangorestframework==3.14.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.7
Faker==6.0.0
pytest==6.2.5
pytest-django==4.2.0
factory_boy==3.2.0
